@@ -4,14 +4,15 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "noshow"
-version = "1.2.10"
+version = "3.0.0-alpha.1"
 authors = [
   { name="Ruben Peters", email="r.peters-7@umcutrecht.nl" },
-  { name="Eric Wolters", email="e.j.wolters-4@umcutrecht.nl" }
+  { name="Eric Wolters", email="e.j.wolters-4@umcutrecht.nl" },
+  { name="Welmoed Tjepkema", email="w.w.a.tjepkema@umcutrecht.nl" }
 ]
 description = "No-show prediction model to reduce the amount of no-shows at the clinics"
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 classifiers = [
     "Programming Language :: Python :: 3",
     "Operating System :: OS Independent",
@@ -19,34 +20,47 @@ classifiers = [
 dependencies =[
     "pandas~=2.0",
     "matplotlib~=3.7",
-    "numpy~=1.24",
-    "scikit-learn~=1.3",
-    "pyarrow~=14.0",
+    "numpy>=1.24",
+    "scikit-learn>=1.3",
+    "pyarrow>=14.0",
     "fastapi~=0.109",
-    "dvc~=3.33",
-    "dvclive~=3.41",
     "sqlalchemy~=2.0",
     "python-dotenv~=1.0",
-    "pymssql~=2.2",
-    "streamlit~=1.30",
-    "tomli~=2.0",
+    "pymssql>=2.3.3",
+    "streamlit>=1.30",
     "relplot>=1.0",
     "azure-data-tables>=12.5.0",
     "azure-identity>=1.17.1",
+    "holidays>=0.74",
+    "mlflow>=3.1.0",
+    "rich>=14.0.0",
 ]
 
-[project.optional-dependencies]
-test = ["pytest", "pytest-cov", "pytest-asyncio"]
-dev = ["rsconnect-python", "uvicorn", "nbstripout", "ipykernel"]
+[project.scripts]
+train_no_show = "noshow.train_pipeline:train_pipeline"
 
-[tool]
-rye = { dev-dependencies = [
+[dependency-groups]
+dev = [
     "ipykernel>=6.29.2",
     "pytest>=8.0.1",
     "nbstripout>=0.7.1",
-    "uvicorn~=0.28.0",
-    "pytest-asyncio~=0.23.6",
-] }
+    "uvicorn>=0.28.0",
+    "pytest-asyncio>=0.23.6",
+    "pytest-cov>=5.0.0",
+    "alembic>=1.13.3",
+    "rsconnect-python>=1.24.0",
+    "dvc>=3.60.1",
+]
+docs = [
+    "mkdocs>=1.6.1",
+    "mkdocs-material>=9.6.15",
+]
+lint = [
+    "ruff>=0.7.1",
+]
+
+[tool.uv]
+default-groups = ["dev", "lint"]
 
 [tool.isort]
 profile = "black"
@@ -71,3 +85,6 @@ extend-immutable-calls = [
     "fastapi.params.Depends",
     "Security",
 ]
+
+[tool.pytest.ini_options]
+asyncio_default_fixture_loop_scope = "function"
