@@ -1,4 +1,4 @@
-import os
+import logging
 from datetime import datetime, timedelta
 
 import altair as alt
@@ -8,6 +8,7 @@ import streamlit as st
 from dotenv import load_dotenv
 from sqlalchemy import Date, cast, select
 
+from noshow.config import setup_root_logger
 from noshow.dashboard.connection import init_session
 from noshow.database.models import (
     ApiCallResponse,
@@ -17,23 +18,26 @@ from noshow.database.models import (
 )
 from noshow.preprocessing.utils import add_working_days
 
-load_dotenv()
+load_dotenv(override=True)  # VS Code corrupts the .env file so override
+
+logger = logging.getLogger(__name__)
+setup_root_logger()
 
 date_3_days = add_working_days(datetime.today(), 3)
 
-# Global and env variables
-db_user = os.environ["DB_USER"]
-db_passwd = os.environ["DB_PASSWD"]
-db_host = os.environ["DB_HOST"]
-db_port = os.environ["DB_PORT"]
-db_database = os.environ["DB_DATABASE"]
-
 
 def calc_calling_percentage(
-    call_results_df: pd.DataFrame, call_outcomes: list[str]
+    call_results_df: pd.DataFrame,
+    call_outcomes: list[str],
+    include_absolute: bool = True,
+    include_opt_out: bool = False,
 ) -> str:
     """Calculates the percentage of rows that have `call_outcomes` and
     returns it as a string with two decimal places and a percentage sign.
+
+    Also optionally include the absolute number of rows with `call_outcomes`.
+    If `include_opt_out` is True, the percentage is calculated based on
+    the total number of rows with `call_outcomes` and `opt_out` = 1.
 
     Parameters
     ----------
@@ -47,10 +51,24 @@ def calc_calling_percentage(
     str
         The percentage formatted as a string
     """
-    num_outcomes = len(
-        call_results_df[call_results_df["call_outcome"].isin(call_outcomes)]
-    )
+    if include_opt_out:
+        num_outcomes = len(
+            call_results_df[
+                (call_results_df["call_outcome"].isin(call_outcomes))
+                & (call_results_df["opt_out"] == 1)
+            ]
+        )
+    else:
+        num_outcomes = len(
+            call_results_df[call_results_df["call_outcome"].isin(call_outcomes)]
+        )
+
     total = len(call_results_df)
+    if total == 0:
+        return "0 (0.00%)" if include_absolute else "0.00%"
+
+    if include_absolute:
+        return f"{num_outcomes} ({num_outcomes / total * 100:.2f}%)"
     return f"{num_outcomes / total * 100:.2f}%"
 
 
@@ -69,13 +87,14 @@ def kpi_page():
                 ApiPrediction.clinic_name,
                 ApiCallResponse.call_status,
                 ApiCallResponse.call_outcome,
+                ApiPatient.opt_out,
             )
             .outerjoin(ApiPrediction.callresponse_relation)
             .outerjoin(ApiPrediction.patient_relation)
             .where(
                 (cast(ApiPrediction.start_time, Date) >= date_input[0])
                 & (cast(ApiPrediction.start_time, Date) <= date_input[1])
-                & (ApiPatient.treatment_group == 1)
+                & (ApiPatient.treatment_group >= 1)
             )
         ).all()
 
@@ -91,13 +110,13 @@ def kpi_page():
     call_results_df.loc[call_results_df["call_outcome"] == "Geen", "call_outcome"] = (
         "Onbereikbaar"
     )
+
     call_results_df["color_sort"] = call_results_df["call_outcome"].map(
         {
             "Niet gebeld": 0,
             "Onbereikbaar": 1,
             "Herinnerd": 2,
             "Verzet/Geannuleerd": 3,
-            "Bel me niet": 4,
         }
     )
 
@@ -114,29 +133,55 @@ def kpi_page():
 
     metric_cols = st.columns(5)
     metric_cols[0].metric(
-        "Aantal Hoofdagenda's", call_results_df["clinic_name"].nunique()
+        "Aantal patienten gebeld",
+        calc_calling_percentage(
+            call_results_df,
+            [
+                "Herinnerd",
+                "Verzet/Geannuleerd",
+                "Onbereikbaar",
+                "Voicemail ingesproken",
+            ],
+        ),
+        help="Aantal patienten dat is gebeld, onafhankelijk van of ze bereikbaar waren",
     )
     metric_cols[1].metric(
-        "Aantal patienten herinnerd",
-        len(call_results_df[call_results_df["call_outcome"] == "Herinnerd"]),
+        "Aantal patienten bereikt",
+        calc_calling_percentage(call_results_df, ["Herinnerd", "Verzet/Geannuleerd"]),
+        help=(
+            "Aantal patienten dat is bereikt (herinnerd aan de afspraak "
+            "of afspraak verzet/afgezegd)"
+        ),
     )
     metric_cols[2].metric(
-        "Aantal afspraken verzet of geannuleerd",
-        len(call_results_df[call_results_df["call_outcome"] == "Verzet/Geannuleerd"]),
-    )
-    percentage_reached = calc_calling_percentage(
-        call_results_df, ["Herinnerd", "Verzet/Geannuleerd"]
+        "Aantal patienten herinnerd",
+        calc_calling_percentage(call_results_df, ["Herinnerd"]),
+        help="Aantal patienten dat is herinnerd aan de afspraak",
     )
     metric_cols[3].metric(
-        "Percentage patienten bereikt",
-        percentage_reached,
+        "Aantal afspraken verzet of geannuleerd",
+        calc_calling_percentage(call_results_df, ["Verzet/Geannuleerd"]),
+        help="Aantal patienten dat de afspraak heeft verzet of geannuleerd",
     )
-    percentage_called = calc_calling_percentage(
-        call_results_df, ["Herinnerd", "Verzet/Geannuleerd", "Onbereikbaar"]
-    )
+
     metric_cols[4].metric(
-        "Percentage patienten gebeld",
-        percentage_called,
+        "Aantal patienten met opt-out",
+        calc_calling_percentage(
+            call_results_df,
+            [
+                "Herinnerd",
+                "Verzet/Geannuleerd",
+                "Geen",
+                "Bel me niet",
+                "Voicemail ingesproken",
+                "Onbereikbaar",
+            ],
+            include_opt_out=True,
+        ),
+        help=(
+            "Aantal patiënten die gebeld zijn ('Herinnerd' of 'Verzet/Geannuleerd') en "
+            "hebben aangegeven in de toekomst niet meer gebeld te willen worden."
+        ),
     )
 
     st.write("### Uitkomsten per dag")
@@ -146,23 +191,40 @@ def kpi_page():
         "Let op dat de datum de dag van de afspraak is, de patienten worden "
         "drie dagen voor de afspraak gebeld."
     )
+
+    show_not_called = st.checkbox("Toon niet gebelde patiënten", value=False)
+    call_outcomes_plot = [
+        "Herinnerd",
+        "Verzet/Geannuleerd",
+        "Voicemail ingesproken",
+        "Onbereikbaar",
+    ]
+    if show_not_called:
+        call_outcomes_plot.append("Niet gebeld")
+
+    order_mapping = {k: i for i, k in enumerate(call_outcomes_plot)}
+    call_results_df["order"] = call_results_df["call_outcome"].map(order_mapping)
+
     bar_chart = (
-        alt.Chart(call_results_df)
+        alt.Chart(
+            call_results_df[call_results_df["call_outcome"].isin(call_outcomes_plot)]
+        )
         .mark_bar()
         .encode(
-            x=alt.X("date:T").timeUnit("yearmonthdate"),
+            x=alt.X(
+                "yearmonthdate(date):O",
+                axis=alt.Axis(title="Datum", labelAlign="center", labelAngle=0),
+                bandPosition=0.5,
+            ),
             y="count()",
             color=alt.Color(
                 "call_outcome",
-                sort=[
-                    "Niet gebeld",
-                    "Onbereikbaar",
-                    "Herinnerd",
-                    "Verzet/Geannuleerd",
-                    "Bel me niet",
-                ],
+                scale=alt.Scale(
+                    domain=list(order_mapping.keys()),
+                    range=["#006400", "#90ee90", "#FFB90F", "#CD5C5C", "#7171C6"],
+                ),
             ),
-            order=alt.Order("color_sort", sort="descending"),
+            order=alt.Order("order:Q", sort="ascending"),
         )
     )
     st.altair_chart(bar_chart, use_container_width=True)
@@ -191,7 +253,7 @@ def monitoring_page():
             .where(
                 (cast(ApiRequest.timestamp, Date) >= date_input[0])
                 & (cast(ApiRequest.timestamp, Date) <= date_input[1])
-                & (ApiPatient.treatment_group == 1)
+                & (ApiPatient.treatment_group >= 1)
             )
         ).all()
 
@@ -247,7 +309,7 @@ if __name__ == "__main__":
     st.set_page_config("No-Show Admin Dashboard", page_icon="📈", layout="wide")
     st.title("No-Show Admin Dashboard")
 
-    session_object = init_session(db_user, db_passwd, db_host, db_port, db_database)
+    session_object = init_session()
 
     with st.sidebar:
         date_input = st.date_input(
