@@ -1,65 +1,22 @@
-from datetime import date
+import json
+import logging
+from datetime import date, datetime
 from typing import List
 
 import pandas as pd
 import streamlit as st
-from sqlalchemy import select
+from sqlalchemy import Date, cast, select
 from sqlalchemy.orm import sessionmaker
+from streamlit.runtime.context import StreamlitHeaders
 
-from noshow.database.models import ApiCallResponse, ApiPatient, ApiSensitiveInfo
+from noshow.database.models import (
+    ApiCallResponse,
+    ApiPatient,
+    ApiPrediction,
+    ApiSensitiveInfo,
+)
 
-
-def render_patient_info(
-    Session: sessionmaker,
-    current_response: ApiCallResponse,
-    current_patient: ApiSensitiveInfo,
-    current_patient_nmbr: ApiPatient,
-    call_number_list: List[str],
-) -> None:
-    """Render patient information
-
-    This function is responsible for rendering the patient information on the dashboard.
-
-    Parameters
-    ----------
-    Session : sessionmaker
-        The SQLAlchemy sessionmaker object used for database operations.
-    current_response : ApiCallResponse
-        The current call response object.
-    current_patient : ApiSensitiveInfo
-        The current patient object containing sensitive information.
-    current_patient_nmbr : ApiPatient
-        The current patient object containing the call number.
-    call_number_list : List[str]
-        The list of call number types.
-    """
-    if current_response.call_status == "Niet gebeld":
-        st.button(
-            "Start met bellen patient",
-            on_click=start_calling,
-            args=(
-                Session,
-                current_response,
-            ),
-            type="primary",
-        )
-    else:
-        if current_patient:
-            if current_response.call_status != "Wordt gebeld":
-                st.warning("Deze patient is al gebeld!", icon="⚠️")
-            st.write(f"- Naam: {current_patient.full_name or 'Onbekend'}")
-            st.write(f"- Voornaam: {current_patient.first_name or 'Onbekend'}")
-            st.write(f"- Geboortedatum: {current_patient.birth_date or 'Onbekend'}")
-            st.write(f"- Mobiel: {current_patient.mobile_phone or 'Onbekend'}")
-            st.write(f"- Thuis: {current_patient.home_phone or 'Onbekend'}")
-            st.write(f"- Overig nummer: {current_patient.other_phone or 'Onbekend'}")
-            st.write("")
-            if not current_patient_nmbr.call_number:
-                current_patient_nmbr.call_number = 0
-            call_number_type = call_number_list[current_patient_nmbr.call_number]
-            st.write(f"- Eerder contact ging via: {call_number_type or 'Onbekend'}")
-        else:
-            st.write("Patientgegevens zijn verwijderd.")
+logger = logging.getLogger(__name__)
 
 
 def highlight_row(row: pd.Series) -> List[str]:
@@ -83,15 +40,8 @@ def highlight_row(row: pd.Series) -> List[str]:
         return [""] * len(row)
 
 
-def previous_preds(call_status: str = "Niet gebeld"):
+def previous_preds():
     """Go to previous prediction"""
-    if call_status == "Wordt gebeld":
-        st.error(
-            "Status is 'Wordt gebeld', verander de status voordat je verder gaat.",
-            icon="🛑",
-        )
-        return
-
     if st.session_state["pred_idx"] > 0:
         st.session_state["pred_idx"] -= 1
 
@@ -111,10 +61,13 @@ def start_calling(Session: sessionmaker, call_response: ApiCallResponse):
     None
     """
     call_response.call_status = "Wordt gebeld"
+    st.session_state["being_called"] = True
+    st.session_state["saved"] = False
 
     with Session() as session:
         session.merge(call_response)
         session.commit()
+        logger.info("Calling started")
 
 
 def next_preds(
@@ -122,6 +75,7 @@ def next_preds(
     Session: sessionmaker,
     call_response: ApiCallResponse,
     current_patient: ApiPatient,
+    user_name: str,
 ) -> None:
     """Go to the next prediction and save results
 
@@ -135,60 +89,110 @@ def next_preds(
         call response object that needs to be edited
     current_patient : ApiPatient
         current patient object
+    user_name : str
+        The current user e-mail
     """
-    call_response.call_status = st.session_state.status_input
+    if st.session_state.res_input == "Geen":
+        call_response.call_status = "Niet gebeld"
+    else:
+        call_response.call_status = "Gebeld"
     call_response.call_outcome = st.session_state.res_input
     call_response.remarks = st.session_state.opm_input
+    call_response.timestamp = datetime.now()
+    call_response.user = user_name
     current_patient.call_number = st.session_state.number_input
-
-    if call_response.call_status == "Wordt gebeld":
-        st.error(
-            "Status is 'Word Gebeld', verander de status voordat je verder gaat.",
-            icon="🛑",
-        )
-        return
-    if call_response.call_status == "Gebeld":
-        current_patient.last_call_date = date.today()
-    if call_response.call_outcome == "Bel me niet":
-        current_patient.opt_out = 1
-        call_response.call_status = "Gebeld"
+    current_patient.opt_out = st.session_state.opt_out_checkbox
 
     with Session() as session:
         session.merge(call_response)
         session.merge(current_patient)
         session.commit()
+        logger.info("Call response saved to database")
+
     if st.session_state["pred_idx"] + 1 < list_len:
         st.session_state["pred_idx"] += 1
+    else:
+        st.session_state["saved"] = True
+        st.session_state["being_called"] = False
 
 
-def navigate_patients(
-    list_len: int, navigate_forward: bool = True, call_status: str = "Niet gebeld"
-):
+def navigate_patients(list_len: int, navigate_forward: bool = True):
     """Navigate through patients
 
-    Navigates through the patients list and reset the prediction index
+    Navigates through the patients list and resets the prediction index
 
     Parameters
     ----------
     list_len : int
         Length of patient list
     navigate_forward : bool, optional
-        Whether to navigate forward or backword, by default True
+        Whether to navigate forward or backward, by default True
     """
-    if call_status == "Wordt gebeld":
+
+    if not st.session_state["saved"] and st.session_state["being_called"]:
         st.error(
-            "Status is 'Wordt gebeld', verander de status voordat je verder gaat.",
+            "Klik eerst op 'Opslaan' voordat je verder gaat.",
             icon="🛑",
         )
+        logger.warning("Trying to navigate while the call is still in progress.")
         return
-    elif navigate_forward:
+
+    if navigate_forward:
         if st.session_state["name_idx"] + 1 < list_len:
             st.session_state["name_idx"] += 1
             st.session_state["pred_idx"] = 0
+            st.session_state["being_called"] = False
+            st.session_state["saved"] = False
     else:
         if st.session_state["name_idx"] > 0:
             st.session_state["name_idx"] -= 1
             st.session_state["pred_idx"] = 0
+            st.session_state["being_called"] = False
+            st.session_state["saved"] = False
+
+
+def navigate_uncalled(Session: sessionmaker, patient_ids: list[str]) -> None:
+    """Navigate to the first patient who has not been called today.
+
+    Parameters
+    ----------
+    Session : sessionmaker
+        SQLAlchemy sessionmaker object
+    patient_ids : list[str]
+        List of patient IDs
+    """
+    with Session() as session:
+        query = (
+            select(ApiPrediction.patient_id)
+            .outerjoin(ApiPrediction.callresponse_relation)
+            .where(
+                ApiCallResponse.call_outcome.in_(
+                    [
+                        "Herinnerd",
+                        "Voicemail ingesproken",
+                        "Verzet/Geannuleerd",
+                        "Onbereikbaar",
+                    ]
+                )
+            )
+            .where(cast(ApiCallResponse.timestamp, Date) == cast(date.today(), Date))
+            .where(ApiPrediction.active)
+            .distinct()
+        )
+        called_today = session.execute(query).scalars().all()
+        uncalled_patient = next(
+            (
+                patient_id
+                for patient_id in patient_ids
+                if patient_id not in called_today
+            ),
+            None,
+        )
+        if uncalled_patient:
+            st.session_state["name_idx"] = patient_ids.index(uncalled_patient)
+            st.session_state["pred_idx"] = 0
+        else:
+            st.info("Alle patiënten zijn vandaag al gebeld.", icon="ℹ️")
 
 
 def search_number(
@@ -219,7 +223,31 @@ def search_number(
         if patient_id and patient_id in patient_ids:
             st.session_state["name_idx"] = patient_ids.index(patient_id)
             st.session_state["pred_idx"] = 0
+            logger.info("Successfully found patient with phone number")
         else:
             st.info(
                 "Geen patient gevonden met dit telefoonnummer op deze dag", icon="ℹ️"
             )
+            logger.warning("Patient not found with this phone number on this day")
+
+
+def get_user(headers: StreamlitHeaders) -> str:
+    """Get the user from the streamlit headers
+
+    Parameters
+    ----------
+    header : StreamlitHeaders
+        Streamlit headers object, contains a RStudio-Connect-Credentials header
+        when deployed to PositConnect
+
+    Returns
+    -------
+    str
+        The user e-mail in lowercase
+    """
+    credential_header = headers.get("RStudio-Connect-Credentials")
+    if not credential_header:
+        logger.warning("Rsconnect credentials not found")
+        return "No user"
+    credential_header = json.loads(credential_header)
+    return credential_header.get("user").lower()
