@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from noshow.config import APPOINTMENTS_LAST_DAYS, MINUTES_EARLY_CUTOFF
+from noshow.config import APPOINTMENTS_LAST_DAYS, CLINIC_CONFIG, MINUTES_EARLY_CUTOFF
 from noshow.features.appointment_features import (
     add_appointments_last_days,
     add_appointments_same_day,
@@ -83,7 +83,7 @@ if __name__ == "__main__":
     data_path = Path(__file__).parents[3] / "data" / "raw"
     output_path = Path(__file__).parents[3] / "data" / "processed"
     appointments_df = load_appointment_csv(data_path / "poliafspraken_no_show.csv")
-    appointments_df = process_appointments(appointments_df)
+    appointments_df = process_appointments(appointments_df, CLINIC_CONFIG)
     all_postalcodes = process_postal_codes(data_path / "NL.txt")
     appointments_features = (
         create_features(appointments_df, all_postalcodes)
