@@ -1,45 +1,31 @@
+import logging
+from datetime import date
 from pathlib import Path
 from typing import Dict, List, Union
 
 import pandas as pd
 
+from noshow.api.pydantic_models import Appointment
+from noshow.config import NO_SHOW_CODES, ClinicConfig
 
-def load_appointment_json(input: List[Dict]) -> pd.DataFrame:
-    """Load prediction data from a JSON dictionary
+logger = logging.getLogger(__name__)
+
+
+def load_appointment_pydantic(input: List[Appointment]) -> pd.DataFrame:
+    """Load prediction data from a list of Appointments
 
     Parameters
     ----------
-    input : List[Dict]
-        The input data as a dictionary in the records orientation
-        (every dictionary corresponds to a row)
+    input : List[Appointment]
+        The input data as a list of Appointments
 
     Returns
     -------
     pd.DataFrame
-        The loaded JSON data as pandas dataframe
+        The loaded data as pandas dataframe
     """
-    appointments_df = pd.DataFrame.from_records(input, coerce_float=True)
-
+    appointments_df = pd.DataFrame([a.model_dump() for a in input])
     appointments_df = appointments_df.replace("", None)
-    appointments_df["created"] = pd.to_datetime(
-        appointments_df["created"].astype(float), unit="ms"
-    )
-    appointments_df["start"] = pd.to_datetime(
-        appointments_df["start"].astype(float), unit="ms"
-    )
-    appointments_df["end"] = pd.to_datetime(
-        appointments_df["end"].astype(float), unit="ms"
-    )
-    appointments_df["gearriveerd"] = pd.to_datetime(
-        appointments_df["gearriveerd"].astype(float), unit="ms"
-    )
-    appointments_df["birthDate"] = pd.to_datetime(
-        appointments_df["birthDate"].astype(float), unit="ms"
-    )
-
-    appointments_df["address_postalCodeNumbersNL"] = appointments_df[
-        "address_postalCodeNumbersNL"
-    ].astype(float)
 
     # Clinic name and description are sometimes unknown since HiX6.3
     appointments_df.loc[appointments_df["name"].isnull(), "name"] = "Onbekend"
@@ -66,34 +52,63 @@ def load_appointment_csv(csv_path: Union[str, Path]) -> pd.DataFrame:
     """
     appointments_df = pd.read_csv(
         csv_path,
-        parse_dates=["created"],
+        engine="pyarrow",
     )
+
     return appointments_df
 
 
-def process_appointments(appointments_df: pd.DataFrame) -> pd.DataFrame:
+def process_appointments(
+    appointments_df: pd.DataFrame,
+    clinic_config: Dict[str, ClinicConfig],
+    start_date: str | None = None,
+) -> pd.DataFrame:
     """Process the appointments data
 
     Parameters
     ----------
     appointments_df : Union[str, Path]
         The pandas dataframe with appointments data from either csv or json
+    clinic_config : Dict[str, ClinicConfig]
+        The clinic configuration, containing filters and clinic info
+    start_date : str, optional
+        The start date for the predictions, if given will filter out appointments of
+        patients that do not have an appointment on this date, by default None
 
     Returns
     -------
     pd.DataFrame
         Cleaned appointment DataFrame that can be used for feature building
     """
-    appointments_df["start"] = pd.to_datetime(appointments_df["start"], errors="coerce")
-    appointments_df["end"] = pd.to_datetime(appointments_df["end"], errors="coerce")
-    appointments_df["gearriveerd"] = pd.to_datetime(
-        appointments_df["gearriveerd"], errors="coerce"
-    )
+    appointments_df = apply_config_filters(appointments_df, clinic_config, start_date)
 
     appointments_df["no_show"] = "show"
-    appointments_df.loc[appointments_df["cancelationReason_code"] == "N", "no_show"] = (
-        "no_show"
+    appointments_df.loc[
+        appointments_df["mutationReason_code"].isin(NO_SHOW_CODES), "no_show"
+    ] = "no_show"
+
+    # Appointments that have status finished but mutation No-Show could have
+    # been moved and need to be filtered out
+    no_show_finished_mask = (appointments_df["no_show"] == "no_show") & (
+        appointments_df["status"] == "finished"
     )
+    if no_show_finished_mask.any():
+        appointments_df = appointments_df.loc[~no_show_finished_mask]
+        logger.warning(
+            f"Filtered out {no_show_finished_mask.sum()} no-show "
+            "appointments with status 'finished'."
+        )
+
+    # Remove appointments that are cancelled for other reasons than no-show
+    cancelled_mask = (appointments_df["status"] == "cancelled") & (
+        appointments_df["no_show"] != "no_show"
+    )
+    if cancelled_mask.any():
+        appointments_df = appointments_df.loc[~cancelled_mask]
+        logger.info(
+            f"Filtered out {cancelled_mask.sum()} cancelled appointments "
+            "that are not no-shows."
+        )
 
     # Some patients have multiple postal codes
     appointments_df = appointments_df.drop_duplicates(
@@ -154,3 +169,59 @@ def process_postal_codes(postalcodes_path: Union[str, Path]) -> pd.DataFrame:
     all_postalcodes = all_postalcodes.set_index("postalcode")[["latitude", "longitude"]]
     all_postalcodes = all_postalcodes.loc[~all_postalcodes.index.duplicated()]
     return all_postalcodes
+
+
+def apply_config_filters(
+    appointments_df: pd.DataFrame,
+    clinic_config: Dict[str, ClinicConfig],
+    start_date: str | None = None,
+) -> pd.DataFrame:
+    """Apply the clinic config filters to the appointments data
+
+    Parameters
+    ----------
+    appointments_df : pd.DataFrame
+        The appointments data
+    clinic_config : Dict[str, ClinicConfig]
+        The clinic configuration, containing filters and clinic info
+    start_date : str, optional
+        The start date for the predictions, if given will filter out appointments of
+        patients that do not have an appointment on this date, by default None
+
+    Returns
+    -------
+    pd.DataFrame
+        The appointments data with the clinic config filters applied
+    """
+    clinic_df_list = []
+    for name, config in clinic_config.items():
+        clinic_df = appointments_df.loc[
+            appointments_df["hoofdagenda_id"].isin(config.main_agenda_codes)
+        ].copy()
+        clinic_df["clinic"] = name
+
+        if config.subagenda_exclude and config.subagendas:
+            clinic_df = clinic_df.loc[
+                ~clinic_df["subagenda_id"].isin(config.subagendas)
+            ]
+        elif config.subagendas:
+            clinic_df = clinic_df.loc[clinic_df["subagenda_id"].isin(config.subagendas)]
+
+        if config.appcode_exclude and config.appcodes:
+            clinic_df = clinic_df.loc[~clinic_df["afspraak_code"].isin(config.appcodes)]
+        elif config.appcodes:
+            clinic_df = clinic_df.loc[clinic_df["afspraak_code"].isin(config.appcodes)]
+
+        clinic_df_list.append(clinic_df)
+
+    total_df = pd.concat(clinic_df_list)
+
+    # After filtering there could still be appointments of patients that no longer have
+    # an appointment on the start date.
+    if start_date:
+        patient_list = total_df.loc[
+            total_df["start"].dt.date == date.fromisoformat(start_date), "pseudo_id"
+        ].unique()
+        total_df = total_df.loc[total_df["pseudo_id"].isin(patient_list)]
+
+    return total_df
