@@ -3,19 +3,27 @@ from tempfile import TemporaryDirectory
 
 import pandas as pd
 from sklearn.dummy import DummyClassifier
-from test_noshow import FakeModel, fake_appointments, fake_postal_codes
+from test_noshow import (
+    FakeModel,
+    create_unit_test_clinic_config,
+    fake_appointments,
+    fake_postal_codes,
+)
 
 from noshow.features.feature_pipeline import create_features, select_feature_columns
 from noshow.model.predict import create_prediction
 from noshow.model.train_model import train_cv_model
-from noshow.preprocessing.load_data import process_appointments
+from noshow.preprocessing.load_data import (
+    load_appointment_pydantic,
+    process_appointments,
+)
 
 
 def test_create_prediction():
-    appointments_df = pd.DataFrame(fake_appointments())
-    appointments_df["created"] = pd.to_datetime(appointments_df["created"], unit="ms")
-    appointments_df["start"] = pd.to_datetime(appointments_df["start"], unit="ms")
-    appointments_df = process_appointments(appointments_df)
+    appointments_df = load_appointment_pydantic(fake_appointments())
+    appointments_df = process_appointments(
+        appointments_df, create_unit_test_clinic_config()
+    )
 
     preds = create_prediction(FakeModel(), appointments_df, fake_postal_codes(None))
     assert preds.shape == (12, 1)  # 12 appointments in test data
@@ -33,13 +41,10 @@ def test_create_prediction():
 
 
 def test_train_model():
-    appointments_df = pd.DataFrame(fake_appointments())
-    appointments_df["created"] = pd.to_datetime(appointments_df["created"], unit="ms")
-    appointments_df["start"] = pd.to_datetime(appointments_df["start"], unit="ms")
-    appointments_df["gearriveerd"] = pd.to_datetime(
-        appointments_df["gearriveerd"], unit="ms"
+    appointments_df = load_appointment_pydantic(fake_appointments())
+    appointments_df = process_appointments(
+        appointments_df, create_unit_test_clinic_config()
     )
-    appointments_df = process_appointments(appointments_df)
     feature_table = create_features(appointments_df, fake_postal_codes(None)).pipe(
         select_feature_columns
     )
@@ -57,13 +62,11 @@ def test_train_model():
         ]
     )
     with TemporaryDirectory() as tempdirname:
-        (Path(tempdirname) / "models").mkdir()
         train_cv_model(
             feature_table,
             tempdirname,
             DummyClassifier(),
             param_grid={},
-            save_dvc_exp=False,
-            dvcyaml=None,  # Prevents writing to dvc.yaml
+            save_exp=False,
         )
-        assert (Path(tempdirname) / "models" / "no_show_model_cv.pickle").is_file()
+        assert (Path(tempdirname) / "no_show_model_cv.pickle").is_file()
