@@ -20,6 +20,7 @@ def sample_df():
         "prediction": [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0],
         "hoofdagenda": ["A", "A", "B", "B", "C", "C", "D", "D", "E", "E"],
         "pseudo_id": [1, 2, 3, 1, 3, 2, 4, 1, 2, 4],
+        "clinic": ["C1", "C1", "C2", "C2", "C3", "C3", "C5", "C5", "C5", "C5"],
     }
     return pd.DataFrame(data)
 
@@ -54,62 +55,87 @@ def some_patients():
     return data
 
 
-class TestRCT:
-    def test_quantile_binning(self, sample_df, no_patients, sample_bins):
-        """
-        Test function for the quantile binning process and treatment group assignment.
-        Raises:
-            AssertionError: If there are issues with bin creation,
-                            structure of the result DataFrame,
-                            or treatment group assignment not alternating within groups.
-        """
-        original_columns = set(sample_df.columns)
+def test_quantile_binning(sample_df, no_patients, sample_bins):
+    """
+    Test function for the quantile binning process and treatment group assignment.
+    Raises:
+        AssertionError: If there are issues with bin creation,
+                        structure of the result DataFrame,
+                        or treatment group assignment not alternating within groups.
+    """
+    original_columns = set(sample_df.columns)
 
-        session = Mock()
-        session.query.return_value.filter.return_value.all.return_value = no_patients
+    session = Mock()
+    session.query.return_value.filter.return_value.all.return_value = no_patients
 
-        # Apply the function
-        result = create_treatment_groups(sample_df, session, sample_bins)
+    # Apply the function
+    result = create_treatment_groups(sample_df, session, sample_bins)
 
-        assert set(result.columns) == original_columns.union(
-            {"treatment_group"}
-        ), "Result DataFrame does not have the correct structure"
-        assert (
-            result["treatment_group"].dtype == int
-        ), "treatment_group column is not of type int"
-        # check if every pseudo_id has one unique treatment group
-        assert (
-            result.groupby("pseudo_id")["treatment_group"].nunique().max() == 1
-        ), "Some pseudo_ids have multiple treatment groups"
+    assert set(result.columns) == original_columns.union(
+        {"treatment_group"}
+    ), "Result DataFrame does not have the correct structure"
+    assert (
+        result["treatment_group"].dtype == int
+    ), "treatment_group column is not of type int"
+    # check if every pseudo_id has one unique treatment group
+    assert (
+        result.groupby("pseudo_id")["treatment_group"].nunique().max() == 1
+    ), "Some pseudo_ids have multiple treatment groups"
 
-    def test_empty_df(self, empty_df, no_patients, sample_bins):
-        """
-        Test function for handling an empty DataFrame.
-        """
-        session = Mock()
-        session.query.return_value.filter.return_value.all.return_value = no_patients
-        # if value error is raised, the function is working correctly
-        with pytest.raises(ValueError):
-            result = create_treatment_groups(empty_df, session, sample_bins)
-            assert result
 
-    def test_create_treatment_groups_no_patients(
-        self, sample_df, no_patients, sample_bins
-    ):
-        # Test case for no patients in the database
-        session = Mock()
-        session.query.return_value.filter.return_value.all.return_value = no_patients
-        result = create_treatment_groups(sample_df, session, sample_bins)
-        result = result["treatment_group"].values
-        assert (result == [1, 0, 0, 0, 1, 0, 0, 0, 0, 1]).all()
+def test_empty_df(empty_df, no_patients, sample_bins):
+    """
+    Test function for handling an empty DataFrame.
+    """
+    session = Mock()
+    session.query.return_value.filter.return_value.all.return_value = no_patients
+    # if value error is raised, the function is working correctly
+    with pytest.raises(ValueError):
+        result = create_treatment_groups(empty_df, session, sample_bins)
+        assert result
 
-    def test_create_treatment_groups_with_patients(
-        self, sample_df, some_patients, sample_bins
-    ):
-        # Test case for patients in the database
-        session = Mock()
-        session.query.return_value.filter.return_value.all.return_value = some_patients
-        result = create_treatment_groups(sample_df, session, sample_bins)
 
-        result = result["treatment_group"].values
-        assert (result == [0, 1, 1, 0, 0, 0, 1, 1, 1, 0]).all()
+def test_create_treatment_groups_no_patients(sample_df, no_patients, sample_bins):
+    # Test case for no patients in the database
+    session = Mock()
+    session.query.return_value.filter.return_value.all.return_value = no_patients
+    result = create_treatment_groups(sample_df, session, sample_bins)
+    assert (result["treatment_group"] == [1, 0, 0, 0, 1, 0, 0, 0, 0, 1]).all()
+
+
+def test_create_treatment_groups_with_patients(sample_df, some_patients, sample_bins):
+    # Test case for patients in the database
+    session = Mock()
+    session.query.return_value.filter.return_value.all.return_value = some_patients
+    result = create_treatment_groups(sample_df, session, sample_bins)
+
+    assert (result["treatment_group"] == [0, 1, 1, 0, 0, 0, 1, 1, 1, 0]).all()
+
+
+def test_create_treatment_group_empty_rct_agendas(
+    sample_df, some_patients, sample_bins
+):
+    """Test case for empty rct_agendas, should add all patients to group 2"""
+    session = Mock()
+    session.query.return_value.filter.return_value.all.return_value = some_patients
+    result = create_treatment_groups(sample_df, session, sample_bins, rct_agendas=[])
+
+    result = result["treatment_group"]
+    assert (result == 2).all()
+
+
+def test_create_treatment_group_some_rct_agendas(sample_df, some_patients, sample_bins):
+    """Test case for some rct_agendas, should add patients of A and B to RCT,
+    meaning group 0 or 1 and patients of other agendas to group 2"""
+    session = Mock()
+    session.query.return_value.filter.return_value.all.return_value = some_patients
+    result = create_treatment_groups(
+        sample_df, session, sample_bins, rct_agendas=["C1", "C5"]
+    )
+
+    assert (
+        result.loc[result["clinic"].isin(["C1", "C5"]), "treatment_group"] < 2
+    ).all()
+    assert (
+        result.loc[~result["clinic"].isin(["C1", "C5"]), "treatment_group"] == 2
+    ).all()
